# backend/microservices/chatdashboardbff/contract/openapi.yaml
openapi: 3.1.0
info:
  title: Chatdashboard BFF API
  description: "REST facade for the chat dashboard, backed by accountservice's gRPC\
    \ API"
  version: v1
servers:
- url: http://localhost:8100
  description: Generated server url
tags:
- name: Tenants
  description: "Tenant CRUD, proxied to accountservice over gRPC"
paths:
  /api/v1/tenants/{tenantId}:
    get:
      tags:
      - Tenants
      summary: Get a tenant by id
      operationId: get
      parameters:
      - name: tenantId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TenantResponse"
    put:
      tags:
      - Tenants
      summary: Update a tenant's name/status (slug is immutable)
      operationId: update
      parameters:
      - name: tenantId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenantUpdateRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TenantResponse"
    delete:
      tags:
      - Tenants
      summary: Delete a tenant
      operationId: delete
      parameters:
      - name: tenantId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: No Content
  /api/v1/tenants:
    get:
      tags:
      - Tenants
      summary: List tenants (paginated)
      operationId: list
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 20
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TenantPageResponse"
    post:
      tags:
      - Tenants
      summary: Create a tenant
      operationId: create
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenantRequest"
        required: true
      responses:
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/TenantResponse"
components:
  schemas:
    TenantUpdateRequest:
      type: object
      properties:
        name:
          type: string
          example: Acme Corp
          maxLength: 255
          minLength: 0
        status:
          type: string
          enum:
          - ACTIVE
          - SUSPENDED
      required:
      - name
      - status
    TenantResponse:
      type: object
      properties:
        tenantId:
          type: string
          format: uuid
        slug:
          type: string
        name:
          type: string
        status:
          type: string
          enum:
          - ACTIVE
          - SUSPENDED
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        createdBy:
          type: string
        updatedBy:
          type: string
        version:
          type: integer
          format: int64
    TenantRequest:
      type: object
      properties:
        slug:
          type: string
          description: "Globally unique, lowercase, becomes a NATS subject segment"
          example: acme-corp
          minLength: 1
          pattern: "^[a-z0-9-]{3,32}$"
        name:
          type: string
          example: Acme Corp
          maxLength: 255
          minLength: 0
        status:
          type: string
          description: Defaults to ACTIVE when omitted
          enum:
          - ACTIVE
          - SUSPENDED
      required:
      - name
      - slug
    TenantPageResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: "#/components/schemas/TenantResponse"
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32